Service rename to work with Zeitwerk
https://github.com/seq-code/registry/issues/268
